requirements.txt
Django==3.2.14
djangorestframework==3.13.1
requests==2.28.1
six==1.16.0
django-requestlogs==0.7.1